Check for the static dir only when necessary
We originally checked for the static dir whenever the server was
initialised, even if it was only running for a backup or restore. It
turned out this check is already done per-command in Main anyway, but
just to be careful I moved the check from initialisation to running,
where it's actually relevant.